Fixed script crash for accessing empty ped data
When you access a ped to get their networkID or delegate a task, if you switch from traffic stops fast, the ped will empty itself out and be prone to trigger accessing a null variable or delegating tasks to a null variable.

Created new function "isPedEmpty" that checks for ped being null or networkid=0 (empty), and added it before every task delegation.