Bug Fixes
In this update I corrected an issue where peds would walk away after being cuffed.

There was also another bug fix where ped attacking with a knife was not working.

Last update was adding an animation for whenever the ped pulls a weapon when outside of the car instead of just adding it directly to their hand.