HotFix for unattended LEO vehicles breaking
Added an extra parameter that checks if the vehicle that is being targeted has the driver seat free or not. If for any reason the driver verification fails, this vehicle verification should tell us if there is a driver or not.

If there is no driver, nothing will happen to the vehicle, regardless of being LEO or not.